#11 Split out generator functions to own handler
I honestly have no data on this working as intended.
But I did lower the RAM for that main function to 256, and
it seems to perform about the same.

I suspect that the javascript bundle that is loaded into
the lambda is of the same size, though. So there
is further optimizations to be made.